JumpscaleLibsExtra/tools/storybot/tests/requirements.txt
nose
nose-testconfig
requests